ajustando bug v.6

+++ src/main/java/com/artecomcarinho/dto/CustomerKpiDTO.java
@@ -6,6 +6,6 @@ import java.time.LocalDateTime;
 public record CustomerKpiDTO(
         Long customerId,
         Long ordersCount,
-        Double totalRevenue,
+        BigDecimal totalRevenue,
         LocalDateTime lastOrderDate
 ) {}
